src/airflow/requirements.txt
openai==2.32.0
minio==7.2.5
sqlalchemy
pandas
requests
pymupdf
python-docx